Refactor test utils and improve usePagination with useRef implementation
