+++ src/__tests__/utils/test-utils.tsx
@@ -1,29 +1,18 @@
-import React, { ReactElement } from 'react';
-import { render as rtlRender, RenderOptions } from '@testing-library/react';
+import { ReactElement } from 'react';
+import {
+  render as rtlRender,
+  RenderOptions,
+  screen,
+  fireEvent,
+  waitFor,
+  within,
+  act,
+  cleanup
+} from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import { BrowserRouter } from 'react-router-dom';
-import { ProjectProvider } from '@/context/ProjectContext';
-import { ToastProvider } from '@/context/ToastContext';
-import { ToastContainer } from 'react-toastify';
+import AllProviders from './TestProviders';
 
-// Interface for providers props
-interface ProvidersProps {
-  children: React.ReactNode;
-}
 
-// Add any providers that components need to the wrapper
-function AllProviders({ children }: ProvidersProps): JSX.Element {
-  return (
-    <BrowserRouter>
-      <ProjectProvider>
-        <ToastProvider>
-          {children}
-          <ToastContainer />
-        </ToastProvider>
-      </ProjectProvider>
-    </BrowserRouter>
-  );
-}
 
 // Interface for the extended render result
 interface CustomRenderResult extends ReturnType<typeof rtlRender> {
@@ -44,8 +33,15 @@ function customRender(
   };
 }
 
-// Re-export everything from @testing-library/react
-export * from '@testing-library/react';
+// Export specific functions from @testing-library/react
+export {
+  screen,
+  fireEvent,
+  waitFor,
+  within,
+  act,
+  cleanup
+};
 
 // Override the render method with our customized version
 export { customRender as render };
